==> notebooks/1.2_scrape_oldlistings_all.ipynb <==
# Scrape old listing data All
this is for queued scraping for running it whole night

# this notebook may seems repeated
# but it is for concurrence with the other notebooks

import pickle
base_url = 'https://www.oldlistings.com.au'
# import requests
import grequests
from itertools import islice
from tqdm import tqdm
import time
import sys
sys.path.append('../')
from scripts.env import R_KEY, KEY_1,KEY_2,KEY_3,KEY_4

with open("../data/landing/oldlistings/rent_pages.p", "rb") as f:
    list_rent_2 = pickle.load(f)
with open("../data/landing/oldlistings/buy_pages.p", "rb") as f:
    list_buy_2 = pickle.load(f)

def fetch_in_batches_2(urls, batch_size=5, base_url = 'https://www.oldlistings.com.au'):
    print(f"Fetching {len(urls)/batch_size} batches")
    for batch in tqdm(iter_batches(urls, batch_size)):
        rs = (grequests.get('https://api.scraperapi.com/', params = { 'api_key': scraper_api_key, 'url': base_url + url }) for url in batch)
        responses = grequests.map(rs)
        for i in range(len(responses)):
            _items = batch[i].split('/')
            r = responses[i]
            # print(r)
            with open(f"../data/landing/oldlistings/{_items[-2]}_html/{_items[-3]}_{_items[-4]}_{_items[-1]}.html", "w") as f:
                f.write(r.text)
def iter_batches(iterable, batch_size):
    it = iter(iterable)
    while True:
        batch = list(islice(it, batch_size))
        if not batch:
            break
        yield batch

# func

def is_proper_html(file_path):
    try:
        # Open the file and read the first few lines or bytes (enough to check the start)
        with open(file_path, 'r', encoding='utf-8') as f:
            first_line = f.read(256).strip().lower()  # Read first 256 bytes/characters

        # Check if it starts with common HTML starting tags
        if first_line.startswith('<!doctype html>') or first_line.startswith('<html>'):
            return True
        else:
            return False

    except FileNotFoundError:
        # print(f"File {file_path} not found.")
        return False
    except Exception as e:
        # print(f"An error occurred: {e}")
        return False

def path_generator(link):
    _items = link.split('/')
    return f"../data/landing/oldlistings/{_items[-2]}_html/{_items[-3]}_{_items[-4]}_{_items[-1]}.html"
# # Example usage
# file_path = 'example.html'
# if is_proper_html(file_path):
#     print("The file starts with a proper HTML format.")
# else:
#     print("The file does not start with a proper HTML format.")


# scrape

## 0-5000


# verify the files
count = 0
for link in list_rent_2[0:5000]:
    if is_proper_html(path_generator(link)):
        count += 1
count

scraper_api_key = R_KEY
fetch_in_batches_2([link for link in list_rent_2[5000:5010] if not is_proper_html(path_generator(link))])

## 5000-10000

# time.sleep(10)
lo = 5000
hi = 10000
count = 0
for link in list_rent_2[lo:hi]:
    if is_proper_html(path_generator(link)):
        count += 1
count

scraper_api_key = KEY_4
fetch_in_batches_2([link for link in list_rent_2[lo:hi] if not is_proper_html(path_generator(link))])

## 10000-15000

# time.sleep(10)
lo = 10000
hi = 15000
count = 0
for link in list_rent_2[lo:hi]:
    if is_proper_html(path_generator(link)):
        count += 1
count

scraper_api_key = R_KEY
fetch_in_batches_2([link for link in list_rent_2[lo:hi] if not is_proper_html(path_generator(link))])

## 15000-20000

# time.sleep(10)
lo = 15000
hi = 20000
count = 0
for link in list_rent_2[lo:hi]:
    if is_proper_html(path_generator(link)):
        count += 1
count

scraper_api_key = R_KEY
fetch_in_batches_2([link for link in list_rent_2[lo:hi] if not is_proper_html(path_generator(link))])

## 20000 - 25000

time.sleep(10)
lo = 20000
hi = 25000
count = 0
for link in list_rent_2[lo:hi]:
    if is_proper_html(path_generator(link)):
        count += 1
count

scraper_api_key = R_KEY
fetch_in_batches_2([link for link in list_rent_2[lo:hi] if not is_proper_html(path_generator(link))])

## Random sections

lo = 0
hi = len(list_rent_2)
print(hi)
count = 0
for link in list_rent_2[lo:hi]:
    if is_proper_html(path_generator(link)):
        count += 1
print(count)

scraper_api_key = R_KEY
fetch_in_batches_2([link for link in list_rent_2 if not is_proper_html(path_generator(link))])

import os
# count how many files are there
len(os.listdir('../data/landing/oldlistings/rent_html'))
