==> notebooks/2.1_preprocessing_combine_buy_sell.ipynb <==
import sys, os
sys.path.append(os.path.abspath('../'))
from scripts.utils import create_dir, get_runtime
import time 
start_time = time.time()

import pandas as pd

df0 = pd.read_csv("../data/raw/oldlistings_buy/oldlistings_buy_0.csv")
df1 = pd.read_csv("../data/raw/oldlistings_buy/oldlistings_buy_1.csv")
df2 = pd.read_csv("../data/raw/oldlistings_buy/oldlistings_buy_2.csv")
df3 = pd.read_csv("../data/raw/oldlistings_buy/oldlistings_buy_3.csv")
df4 = pd.read_csv("../data/raw/oldlistings_buy/oldlistings_buy_4.csv")
df5 = pd.read_csv("../data/raw/oldlistings_buy/oldlistings_buy_5.csv")
df6 = pd.read_csv("../data/raw/oldlistings_buy/oldlistings_buy_6.csv")

# combine all dataframes
df = pd.concat([df0, df1, df2, df3, df4, df5, df6])

df

translation_table = {
    "Α": "A", "Β": "B", "Ε": "E", "Ζ": "Z", "Η": "H", "Ι": "I", "Κ": "K", "Μ": "M", "Ν": "N", 
    "Ο": "O", "Ρ": "P", "Τ": "T", "Υ": "Y", "Χ": "X", "О": "O", "Ս": "U", "Ε": "E", "В": "B", 
    "Α": "A", "С": "C", "Ϳ": "J", "Ԍ": "G", "Ε": "E"
}

# Function to clean "address"
df['address_cleaned'] = (
    df['address_cleaned']
    .str.replace("Α", "A")
    .str.replace("Β", "B")
    .str.replace("Γ", "G")
    .str.replace("Δ", "D")
    .str.replace("Ε", "E")
    .str.replace("Ζ", "Z")
    .str.replace("Η", "H")
    .str.replace("Θ", "Th")
    .str.replace("Ι", "I")
    .str.replace("Κ", "K")
    .str.replace("Λ", "L")
    .str.replace("Μ", "M")
    .str.replace("Ν", "N")
    .str.replace("Ξ", "X")
    .str.replace("Ο", "O")
    .str.replace("Π", "P")
    .str.replace("Ρ", "R")  
    .str.replace("Σ", "S")
    .str.replace("Τ", "T")
    .str.replace("Υ", "Y")
    .str.replace("Φ", "F")
    .str.replace("Χ", "Ch")
    .str.replace("Ψ", "Ps")
    .str.replace("Ω", "W")
    .str.replace("О", "O")
    .str.replace("Ս", "U")
    .str.replace("Ε", "E")
    .str.replace("В", "B")
    .str.replace("Α", "A")
    .str.replace("С", "C")
    .str.replace("Ϳ", "J")
    .str.replace("Ԍ", "G")
    .str.replace("Ε", "E")
)

# Function to clean "suburb"
df['suburb'] = (
    df['suburb']
    .str.replace("Α", "A")
    .str.replace("Β", "B")
    .str.replace("Γ", "G")
    .str.replace("Δ", "D")
    .str.replace("Ε", "E")
    .str.replace("Ζ", "Z")
    .str.replace("Η", "H")
    .str.replace("Θ", "Th")
    .str.replace("Ι", "I")
    .str.replace("Κ", "K")
    .str.replace("Λ", "L")
    .str.replace("Μ", "M")
    .str.replace("Ν", "N")
    .str.replace("Ξ", "X")
    .str.replace("Ο", "O")
    .str.replace("Π", "P")
    .str.replace("Ρ", "R")  
    .str.replace("Σ", "S")
    .str.replace("Τ", "T")
    .str.replace("Υ", "Y")
    .str.replace("Φ", "F")
    .str.replace("Χ", "Ch")
    .str.replace("Ψ", "Ps")
    .str.replace("Ω", "W")
    .str.replace("О", "O")
    .str.replace("Ս", "U")
    .str.replace("Ε", "E")
    .str.replace("В", "B")
    .str.replace("Α", "A")
    .str.replace("С", "C")
    .str.replace("Ϳ", "J")
    .str.replace("Ԍ", "G")
    .str.replace("Ε", "E")
)

create_dir("../data/raw/oldlistings_buy/")
df.to_csv("../data/raw/oldlistings_buy/see.csv", index=False)

# Calculate average property price by year, suburb
df_avg = df.groupby(['year', 'suburb']).agg(
    avg_property_price=('property_price_cleaned', 'mean'),
).reset_index()
df_avg

# Remove the leading and trailing whitespaces from the suburb column
df_avg['suburb'] = df_avg['suburb'].str.strip()

# convert 'year' to int
df_avg['year'] = df_avg['year'].astype(int)

df_avg

create_dir("../data/raw/oldlistings_buy/")
df_avg.to_csv("../data/raw/oldlistings_buy/oldlistings_buy_avg.csv", index=False)
