==> notebooks/2.4_preprocessing_ptv.ipynb <==
# Preprocessing PTV
- This notebook preprocesses the PTV data.
- The preprocessed datasets are exported into the `./data/raw` directory.

import sys, os
sys.path.append(os.path.abspath('../'))
from scripts.utils import create_dir, get_runtime
import time 
start_time = time.time()

import pandas as pd
import geopandas as gpd
from shapely.geometry import Point

## Stops

### 1 - Regional Train

stop_1 = pd.read_csv("../data/landing/ptv/1/1/stops.txt")
stop_1

stop_1_filtered = stop_1[['stop_name', 'stop_lat', 'stop_lon']]
stop_1_filtered

# Create a geometry column based on the latitude and longitude
geometry = [Point(xy) for xy in zip(stop_1_filtered['stop_lon'], stop_1_filtered['stop_lat'])]

# Create a GeoDataFrame
stop_1_gdf = gpd.GeoDataFrame(stop_1_filtered, geometry=geometry)

# Set a coordinate reference system (CRS), for example, WGS84 (EPSG:4326)
stop_1_gdf.set_crs(epsg=4326, inplace=True)

# Display the GeoDataFrame
stop_1_gdf

stops_1_sa2 = gpd.sjoin(stop_1_gdf, SA2, how="left", predicate="within")

# Display the result
stops_1_sa2

create_dir('../data/raw/PTV/1 - Regional Train')
stops_1_sa2.to_csv('../data/raw/PTV/1 - Regional Train/stops_1_sa2.csv', index=False)

### 2 - Metropolitan Train

stop_2 = pd.read_csv("../data/landing/ptv/2/2/stops.txt")
stop_2

stop_2_filtered = stop_2[['stop_name', 'stop_lat', 'stop_lon']]
stop_2_filtered

# Create a geometry column based on the latitude and longitude
geometry = [Point(xy) for xy in zip(stop_2_filtered['stop_lon'], stop_2_filtered['stop_lat'])]

# Create a GeoDataFrame
stop_2_gdf = gpd.GeoDataFrame(stop_2_filtered, geometry=geometry)

# Set a coordinate reference system (CRS), for example, WGS84 (EPSG:4326)
stop_2_gdf.set_crs(epsg=4326, inplace=True)

# Display the GeoDataFrame
stop_2_gdf

SA2

SA2 = gpd.read_file("../data/landing/sa2/sa2-21-shp/SA2_2021_AUST_GDA2020.shp")
SA2 = SA2[["SA2_CODE21", "SA2_NAME21", "geometry"]]
SA2.crs = 'EPSG: 4326'

stops_2_sa2 = gpd.sjoin(stop_2_gdf, SA2, how="left", predicate="within")

# Display the result
stops_2_sa2

#### Push Stops_2_sa2 to `raw` layer

create_dir('../data/raw/PTV/2 - Metropolitan Train')
stops_2_sa2.to_csv('../data/raw/PTV/2 - Metropolitan Train/stops_2_sa2.csv', index=False)

### 3 - Metropolitan Tram

stop_3 = pd.read_csv("../data/landing/ptv/3/3/stops.txt")
# stop_3

stop_3_filtered = stop_3[['stop_name', 'stop_lat', 'stop_lon']]
stop_3_filtered

# Create a geometry column based on the latitude and longitude
geometry = [Point(xy) for xy in zip(stop_3_filtered['stop_lon'], stop_3_filtered['stop_lat'])]

# Create a GeoDataFrame
stop_3_gdf = gpd.GeoDataFrame(stop_3_filtered, geometry=geometry)

# Set a coordinate reference system (CRS), for example, WGS84 (EPSG:4326)
stop_3_gdf.set_crs(epsg=4326, inplace=True)

# Display the GeoDataFrame
stop_3_gdf

stops_3_sa2 = gpd.sjoin(stop_3_gdf, SA2, how="left", predicate="within")

# Display the result
stops_3_sa2

#### Push Stops_3_sa2 to `raw` layer

create_dir('../data/raw/PTV/3 - Metropolitan Tram')
stops_3_sa2.to_csv('../data/raw/PTV/3 - Metropolitan Tram/stops_3_sa2.csv', index=False)

### 4 - Metropolitan Bus

stop_4 = pd.read_csv("../data/landing/ptv/4/4/stops.txt")
# stop_4

stop_4_filtered = stop_4[['stop_name', 'stop_lat', 'stop_lon']]
stop_4_filtered

# Create a geometry column based on the latitude and longitude
geometry = [Point(xy) for xy in zip(stop_4_filtered['stop_lon'], stop_4_filtered['stop_lat'])]

# Create a GeoDataFrame
stop_4_gdf = gpd.GeoDataFrame(stop_4_filtered, geometry=geometry)

# Set a coordinate reference system (CRS), for example, WGS84 (EPSG:4326)
stop_4_gdf.set_crs(epsg=4326, inplace=True)

# Display the GeoDataFrame
stop_4_gdf

stops_4_sa2 = gpd.sjoin(stop_4_gdf, SA2, how="left", predicate="within")

# Display the result
stops_4_sa2

#### Push Stops_4_sa2 to `raw` layer

create_dir('../data/raw/PTV/4 - Metropolitan Bus')
stops_4_sa2.to_csv('../data/raw/PTV/4 - Metropolitan Bus/stops_4_sa2.csv', index=False)

### 5 - Regional Coach

stop_5 = pd.read_csv("../data/landing/ptv/5/5/stops.txt")
# stop_5

stop_5_filtered = stop_5[['stop_name', 'stop_lat', 'stop_lon']]
stop_5_filtered

# Create a geometry column based on the latitude and longitude
geometry = [Point(xy) for xy in zip(stop_5_filtered['stop_lon'], stop_5_filtered['stop_lat'])]

# Create a GeoDataFrame
stop_5_gdf = gpd.GeoDataFrame(stop_5_filtered, geometry=geometry)

# Set a coordinate reference system (CRS), for example, WGS84 (EPSG:4326)
stop_5_gdf.set_crs(epsg=4326, inplace=True)

# Display the GeoDataFrame
stop_5_gdf

stops_5_sa2 = gpd.sjoin(stop_5_gdf, SA2, how="left", predicate="within")

# Display the result
stops_5_sa2

create_dir('../data/raw/PTV/5 - Regional Coach')
stops_5_sa2.to_csv('../data/raw/PTV/5 - Regional Coach/stops_5_sa2.csv', index=False)

### 6 - Regional Bus

stop_6 = pd.read_csv("../data/landing/ptv/6/6/stops.txt")
# stop_6

stop_6_filtered = stop_6[['stop_name', 'stop_lat', 'stop_lon']]
stop_6_filtered

# Create a geometry column based on the latitude and longitude
geometry = [Point(xy) for xy in zip(stop_6_filtered['stop_lon'], stop_6_filtered['stop_lat'])]

# Create a GeoDataFrame
stop_6_gdf = gpd.GeoDataFrame(stop_6_filtered, geometry=geometry)

# Set a coordinate reference system (CRS), for example, WGS84 (EPSG:4326)
stop_6_gdf.set_crs(epsg=4326, inplace=True)

# Display the GeoDataFrame
stop_6_gdf

stops_6_sa2 = gpd.sjoin(stop_6_gdf, SA2, how="left", predicate="within")

# Display the result
stops_6_sa2

create_dir('../data/raw/PTV/6 - Regional Bus')
stops_6_sa2.to_csv('../data/raw/PTV/6 - Regional Bus/stops_6_sa2.csv', index=False)
