==> notebooks/2.7_preprocessing_cpi.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns


# Import all datas

unemployment_data = pd.read_csv('..\\data\\landing\\Unemployment_rate.csv')
cpi_data = pd.read_csv('..\\data\\landing\\cpidata.csv')
housing_interest = pd.read_excel('..\\data\\landing\\Housing_interest_rate.xlsx')

housing_interest

unemployment_data.dropna(inplace=True)
unemployment_data

Extract time columns

unemployment_data['year'] = unemployment_data['Time'].str.split('-').str[1]
unemployment_data['month'] = unemployment_data['Time'].str.split('-').str[0]

# change the year to 20xx
unemployment_data['year'] = '20' + unemployment_data['year']

# change month to number
unemployment_data['month'] = unemployment_data['month'].replace('Jan', '01')
unemployment_data['month'] = unemployment_data['month'].replace('Feb', '02')
unemployment_data['month'] = unemployment_data['month'].replace('Mar', '03')
unemployment_data['month'] = unemployment_data['month'].replace('Apr', '04')
unemployment_data['month'] = unemployment_data['month'].replace('May', '05')
unemployment_data['month'] = unemployment_data['month'].replace('Jun', '06')
unemployment_data['month'] = unemployment_data['month'].replace('Jul', '07')
unemployment_data['month'] = unemployment_data['month'].replace('Aug', '08')
unemployment_data['month'] = unemployment_data['month'].replace('Sep', '09')
unemployment_data['month'] = unemployment_data['month'].replace('Oct', '10')
unemployment_data['month'] = unemployment_data['month'].replace('Nov', '11')
unemployment_data['month'] = unemployment_data['month'].replace('Dec', '12')

unemployment_data

yearly_avg = unemployment_data.groupby('year')['Seasonally adjusted (%)'].mean().reset_index()

yearly_avg.columns = ['year', 'Yearly Average Seasonally Adjusted (%)']

Export the data into csv 

yearly_avg
yearly_avg.to_csv('..\\data\\curated\\yearly_avg_unemployment.csv', index=False)

Process cpi data

cpi_data

# only consider the time when it's december

cpi_data = cpi_data[cpi_data['time'].str.contains('Dec')]

# convert the time to year, which get rid of the month and put 20 in the front
cpi_data["year"] = "20"+cpi_data["time"].str.split('-').str[1]

cpi_data.to_csv('..\\data\\curated\\cpi_data.csv', index=False)


Process housing interest data

housing_interest

housing_interest = housing_interest["Unnamed: 25"]

# the data for housing interest is limited to this type :'Lending rates; Housing credit; Outstanding; Investment; Variable-rate; All institutions'

time = housing_interest['F6 - HOUSING LENDING RATES'][10:]

interest_rate = housing_interest["Unnamed: 25"][10:]

housing_interest_data = pd.DataFrame(columns=['time', 'interest_rate'])
housing_interest_data["time"] = time
housing_interest_data["interest_rate"] = interest_rate

housing_interest_data.to_csv('..\\data\\curated\\housing_interest.csv', index=False)
